refactor(workspace): 🔧 streamline imports and improve workspace initialization logic
Consolidate import paths for workspace-related modules and enhance the workspace initialization process for better maintainability.
