--- client/apps/web/src/workspace/composables/useWorkspaceInitialization.test.ts
@@ -1,8 +1,0 @@
-import { describe, it } from "vitest";
-
-// TODO: Add tests for useWorkspaceInitialization composable
-describe.skip("useWorkspaceInitialization", () => {
-	it("should initialize workspace properly", () => {
-		// Test implementation needed
-	});
-});
