--- client/apps/web/src/workspace/providers/__tests__/workspaceStoreProvider.test.ts
@@ -1,8 +1,0 @@
-import { describe, it } from "vitest";
-
-// TODO: Add tests for workspaceStoreProvider
-describe.skip("workspaceStoreProvider", () => {
-	it("should provide store correctly", () => {
-		// Test implementation needed
-	});
-});
