+++ docs/src/content/docs/structure.md
@@ -6,7 +6,7 @@ description: Overview of the project structure and organization.
 # Project Structure - Overview
 
 <!-- markdownlint-disable-next-line MD036 -->
-**Auto-generated by scripts/generate-structure-docs.sh on 2025-08-11**
+**Auto-generated by scripts/generate-structure-docs.sh on 2025-08-13**
 
 > **Note:** This structure is automatically generated. Do not edit this file manually.
 
@@ -1191,8 +1191,6 @@ description: Overview of the project structure and organization.
 │   │       │   │   │   │   ├───WorkspaceApplicationService.ts
 │   │       │   │   │   │   └───index.ts
 │   │       │   │   │   └───index.ts
-│   │       │   │   ├───composables/
-│   │       │   │   │   └───useWorkspaceInitialization.test.ts
 │   │       │   │   ├───domain/
 │   │       │   │   │   ├───common/
 │   │       │   │   │   │   └───Result.ts
@@ -1240,9 +1238,6 @@ description: Overview of the project structure and organization.
 │   │       │   │   │   ├───validation/
 │   │       │   │   │   │   └───ResponseValidator.ts
 │   │       │   │   │   └───index.ts
-│   │       │   │   ├───providers/
-│   │       │   │   │   └───__tests__/
-│   │       │   │   │       └───workspaceStoreProvider.test.ts
 │   │       │   │   ├───di.ts
 │   │       │   │   └───index.ts
 │   │       │   ├───App.vue
@@ -1609,6 +1604,7 @@ description: Overview of the project structure and organization.
 │   │   │   │   ├───en_7e63627.pf_fragment
 │   │   │   │   ├───en_82142d3.pf_fragment
 │   │   │   │   ├───en_83b65b8.pf_fragment
+│   │   │   │   ├───en_867e224.pf_fragment
 │   │   │   │   ├───en_873d782.pf_fragment
 │   │   │   │   ├───en_8c2c94f.pf_fragment
 │   │   │   │   ├───en_8f30321.pf_fragment
@@ -1617,7 +1613,6 @@ description: Overview of the project structure and organization.
 │   │   │   │   ├───en_9e43836.pf_fragment
 │   │   │   │   ├───en_a1a95ab.pf_fragment
 │   │   │   │   ├───en_a7599ed.pf_fragment
-│   │   │   │   ├───en_a992992.pf_fragment
 │   │   │   │   ├───en_ad47cd6.pf_fragment
 │   │   │   │   ├───en_b3c2234.pf_fragment
 │   │   │   │   ├───en_b46a4b3.pf_fragment
@@ -1646,18 +1641,18 @@ description: Overview of the project structure and organization.
 │   │   │   │   ├───en_faa6dce.pf_fragment
 │   │   │   │   └───en_fbfde01.pf_fragment
 │   │   │   ├───index/
-│   │   │   │   ├───en_375df8f.pf_index
-│   │   │   │   ├───en_56e7556.pf_index
-│   │   │   │   ├───en_6e609cd.pf_index
-│   │   │   │   ├───en_92429f7.pf_index
-│   │   │   │   └───en_a9e83b4.pf_index
+│   │   │   │   ├───en_25e89c7.pf_index
+│   │   │   │   ├───en_59bb4c5.pf_index
+│   │   │   │   ├───en_ca56a64.pf_index
+│   │   │   │   ├───en_d959477.pf_index
+│   │   │   │   └───en_f378944.pf_index
 │   │   │   ├───pagefind-entry.json
 │   │   │   ├───pagefind-highlight.js
 │   │   │   ├───pagefind-modular-ui.css
 │   │   │   ├───pagefind-modular-ui.js
 │   │   │   ├───pagefind-ui.css
 │   │   │   ├───pagefind-ui.js
-│   │   │   ├───pagefind.en_93d2f77f88.pf_meta
+│   │   │   ├───pagefind.en_e7f05c5bcd.pf_meta
 │   │   │   ├───pagefind.js
 │   │   │   ├───wasm.en.pagefind
 │   │   │   └───wasm.unknown.pagefind
@@ -2313,7 +2308,9 @@ description: Overview of the project structure and organization.
 │       │   ├───test-results/
 │       │   │   └───test/
 │       │   │       ├───binary/
-│       │   │       │   └───output.bin
+│       │   │       │   ├───output.bin
+│       │   │       │   ├───output.bin.idx
+│       │   │       │   └───results.bin
 │       │   │       ├───TEST-com.hatchgrid.ApplicationStartupTracesTest.xml
 │       │   │       ├───TEST-com.hatchgrid.ArchTest.xml
 │       │   │       ├───TEST-com.hatchgrid.thryve.BeanValidationTest.xml
